feat: working wasm fibonacci rust program proof

+++ vm/src/system/memory/manager/dimensions.rs
@@ -21,10 +21,10 @@ impl MemoryDimensions {
         self.as_height + self.address_height
     }
     /// Convert an address label (address space, block id) to its index in the memory merkle tree.
-    pub fn label_to_index<F: PrimeField32>(&self, label: (F, usize)) -> usize {
+    pub fn label_to_index<F: PrimeField32>(&self, label: (F, usize)) -> u64 {
         let (addr_space, block_id) = label;
-        ((addr_space.as_canonical_u32() as usize - self.as_offset) << self.address_height)
-            + block_id
+        ((addr_space.as_canonical_u32() as u64 - self.as_offset as u64) << self.address_height)
+            + (block_id as u64)
     }
 }
