+++ vm/src/system/memory/tree/mod.rs
@@ -66,12 +66,12 @@ impl<const CHUNK: usize, F: PrimeField32> MemoryNode<CHUNK, F> {
     }
 
     fn from_memory(
-        memory: &BTreeMap<usize, [F; CHUNK]>,
+        memory: &BTreeMap<u64, [F; CHUNK]>,
         height: usize,
-        from: usize,
+        from: u64,
         hasher: &impl Hasher<CHUNK, F>,
     ) -> MemoryNode<CHUNK, F> {
-        let mut range = memory.range(from..from + (1 << height));
+        let mut range = memory.range(from..from + (1u64 << height));
         if height == 0 {
             let values = *memory.get(&from).unwrap_or(&[F::ZERO; CHUNK]);
             MemoryNode::new_leaf(hasher.hash(&values))
